refactor into a new structure
