+++ cloud/handler/handler.go
@@ -1,7 +1,7 @@
-// Package export is used to import data from external sources
+// Package handler is used to import data from external sources
 // this package has an Import(ctx *gofr.Context) method that is used to import all
 // local cloud accounts to the zop api to store and validate those cloud accounts.
-package export
+package handler
 
 import (
 	"gofr.dev/pkg/gofr"
