+++ cloud/store/gcp/models.go
@@ -1,0 +1,7 @@
+package gcp
+
+// AccountStore stores the accountId and creds value for gcp account.
+type AccountStore struct {
+	AccountID string `json:"account_id"`
+	Value     []byte `json:"value"`
+}
