+++ go.mod
@@ -1,9 +1,10 @@
 module zop.dev/cli/zop
 
-go 1.23.3
+go 1.22.8
 
 require (
 	github.com/mattn/go-sqlite3 v1.14.24
+	github.com/pkg/errors v0.9.1
 	go.uber.org/mock v0.5.0
 	gofr.dev v1.28.0
 	golang.org/x/oauth2 v0.24.0
@@ -51,7 +52,6 @@ require (
 	github.com/ncruces/go-strftime v0.1.9 // indirect
 	github.com/openzipkin/zipkin-go v0.4.3 // indirect
 	github.com/pierrec/lz4/v4 v4.1.21 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/prometheus/client_golang v1.20.5 // indirect
 	github.com/prometheus/client_model v0.6.1 // indirect
